[paramz] started to pull paramz out
