+++ GPy/core/__init__.py
@@ -1,7 +1,7 @@
 # Copyright (c) 2012-2014, GPy authors (see AUTHORS.txt).
 # Licensed under the BSD 3-clause license (see LICENSE.txt)
 
-from .model import *
+from GPy.core.probabilistic_model import *
 from .parameterization.parameterized import adjust_name_for_printing, Parameterizable
 from .parameterization.param import Param, ParamConcatenation
 from .parameterization.observable_array import ObsAr
