+++ GPy/core/gp.py
@@ -4,7 +4,7 @@
 import numpy as np
 import sys
 from .. import kern
-from .model import Model
+from GPy.core.probabilistic_model import Model
 from .parameterization import ObsAr
 from .mapping import Mapping
 from .. import likelihoods
@@ -542,7 +542,7 @@ class GP(Model):
         :param optimize: whether to optimize the location of new X (True by default)
         :type optimize: boolean
         :return: a tuple containing the posterior estimation of X and the model that optimize X
-        :rtype: (:class:`~GPy.core.parameterization.variational.VariationalPosterior` and numpy.ndarray, :class:`~GPy.core.model.Model`)
+        :rtype: (:class:`~GPy.core.parameterization.variational.VariationalPosterior` and numpy.ndarray, :class:`~GPy.core.probabilistic_model.Model`)
         """
         from ..inference.latent_function_inference.inferenceX import infer_newX
         return infer_newX(self, Y_new, optimize=optimize)
