--- GPy/core/parameterization/__init__.py
@@ -1,5 +1,0 @@
-# Copyright (c) 2012, GPy authors (see AUTHORS.txt).
-# Licensed under the BSD 3-clause license (see LICENSE.txt)
-
-from .param import Param, ObsAr
-from .parameterized import Parameterized
