+++ GPy/core/probabilistic_model.py
@@ -1,0 +1,62 @@
+# Copyright (c) 2012-2014, GPy authors (see AUTHORS.txt).
+# Licensed under the BSD 3-clause license (see LICENSE.txt)
+
+
+from .. import likelihoods
+from ..inference import optimization
+from ..util.misc import opt_wrapper
+from .parameterization import Parameterized
+import multiprocessing as mp
+import numpy as np
+from numpy.linalg.linalg import LinAlgError
+import itertools
+import sys
+from .verbose_optimization import VerboseOptimization
+# import numdifftools as ndt
+from functools import reduce
+from paramz.model import Model
+
+class ProbabilisticModel(Model):
+
+    def __init__(self, name):
+        super(ProbabilisticModel, self).__init__(name)  # Parameterized.__init__(self)
+
+    def log_likelihood(self):
+        raise NotImplementedError("this needs to be implemented to use the model class")
+
+    def _log_likelihood_gradients(self):
+        return self.gradient.copy()
+
+
+    def objective_function(self):
+        """
+        The objective function for the given algorithm.
+
+        This function is the true objective, which wants to be minimized.
+        Note that all parameters are already set and in place, so you just need
+        to return the objective function here.
+
+        For probabilistic models this is the negative log_likelihood
+        (including the MAP prior), so we return it here. If your model is not
+        probabilistic, just return your objective to minimize here!
+        """
+        return -float(self.log_likelihood()) - self.log_prior()
+
+    def objective_function_gradients(self):
+        """
+        The gradients for the objective function for the given algorithm.
+        The gradients are w.r.t. the *negative* objective function, as
+        this framework works with *negative* log-likelihoods as a default.
+
+        You can find the gradient for the parameters in self.gradient at all times.
+        This is the place, where gradients get stored for parameters.
+
+        This function is the true objective, which wants to be minimized.
+        Note that all parameters are already set and in place, so you just need
+        to return the gradient here.
+
+        For probabilistic models this is the gradient of the negative log_likelihood
+        (including the MAP prior), so we return it here. If your model is not
+        probabilistic, just return your *negative* gradient here!
+        """
+        return -(self._log_likelihood_gradients() + self._log_prior_gradients())
