+++ GPy/core/variational.py
