+++ GPy/models/gradient_checker.py
@@ -1,7 +1,7 @@
 # ## Copyright (c) 2012, GPy authors (see AUTHORS.txt).
 # Licensed under the BSD 3-clause license (see LICENSE.txt)
 
-from ..core.model import Model
+from GPy.core.probabilistic_model import Model
 import itertools
 import numpy
 from ..core.parameterization import Param
